Remove compiler warnings.

+++ src/MuseRecord.cpp
@@ -112,7 +112,7 @@ string MuseRecord::getOctaveString(void) {
    while ((index < (int)recordInfo.size()) && !std::isdigit(recordInfo[index])) {
       index++;
    }
-   if (index >= recordInfo.size()) {
+   if (index >= (int)recordInfo.size()) {
       cerr << "Error: no octave specification in note field: " << recordInfo
            << endl;
       exit(1);
@@ -161,7 +161,7 @@ int MuseRecord::getPitchClass(void) {
 string MuseRecord::getPitchClassString(void) {
 	string output = getNoteField();
    int index = 0;
-   while ((index < output.size()) &&  !std::isdigit(output[index])) {
+   while ((index < (int)output.size()) &&  !std::isdigit(output[index])) {
       index++;
    }
    output.resize(index);
@@ -180,7 +180,7 @@ int MuseRecord::getAccidental(void) {
    string recordInfo = getNoteField();
    int output = 0;
    int index = 0;
-   while ((index < recordInfo.size()) && (index < 16)) {
+   while ((index < (int)recordInfo.size()) && (index < 16)) {
       if (recordInfo[index] == 'f') {
          output--;
       } else if (recordInfo[index] == '#') {
@@ -3137,7 +3137,7 @@ string MuseRecord::getFigure(int index) {
 	HumRegex hre;
 	vector<string> pieces;
 	hre.split(pieces, temp, " +");
-	if (index < pieces.size()) {
+	if (index < (int)pieces.size()) {
    	output = pieces[index];
 	}
    return output;
